+++ plugins/polio/js/src/domains/VaccineModule/StockManagement/VaccineStockManagement.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from 'react';
 import { useSafeIntl } from 'bluesquare-components';
 import { Box } from '@mui/material';
 import { DisplayIfUserHasPerm } from '../../../../../../../hat/assets/js/apps/Iaso/components/DisplayIfUserHasPerm';
-import { POLIO_VACCINE_STOCK_WRITE } from '../../../../../../../hat/assets/js/apps/Iaso/utils/permissions';
+import { POLIO_VACCINE_STOCK_WRITE, POLIO_VACCINE_STOCK_READ } from '../../../../../../../hat/assets/js/apps/Iaso/utils/permissions';
 import TopBar from '../../../../../../../hat/assets/js/apps/Iaso/components/nav/TopBarComponent';
 import { useStyles } from '../../../styles/theme';
 import MESSAGES from './messages';
@@ -28,7 +28,7 @@ export const VaccineStockManagement: FunctionComponent = () => {
             />
             <Box className={classes.containerFullHeightNoTabPadded}>
                 <VaccineStockManagementFilters params={params} />
-                <DisplayIfUserHasPerm permissions={[POLIO_VACCINE_STOCK_WRITE]}>
+                <DisplayIfUserHasPerm permissions={[POLIO_VACCINE_STOCK_WRITE, POLIO_VACCINE_STOCK_READ]}>
                     <Box mt={2} justifyContent="flex-end" display="flex">
                         <CreateVaccineStock iconProps={{}} />
                     </Box>
